Move peptide search actions
