+++ ms2/src/org/labkey/ms2/query/ProteinGroupTableInfo.java
@@ -29,6 +29,7 @@ import org.labkey.api.data.SQLFragment;
 import org.labkey.api.data.SimpleFilter;
 import org.labkey.api.data.TableInfo;
 import org.labkey.api.data.dialect.SqlDialect;
+import org.labkey.api.protein.MatchCriteria;
 import org.labkey.api.protein.ProteinSchema;
 import org.labkey.api.protein.query.SequencesTableInfo;
 import org.labkey.api.query.ExprColumn;
@@ -40,7 +41,6 @@ import org.labkey.api.view.ActionURL;
 import org.labkey.api.view.ViewContext;
 import org.labkey.ms2.GroupNumberDisplayColumn;
 import org.labkey.ms2.MS2Controller;
-import org.labkey.ms2.MS2Controller.MatchCriteria;
 import org.labkey.ms2.MS2Manager;
 import org.labkey.ms2.MS2Run;
 import org.labkey.ms2.ProteinListDisplayColumn;
@@ -52,10 +52,6 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-/**
- * User: jeckels
- * Date: Feb 8, 2007
- */
 public class ProteinGroupTableInfo extends FilteredTable<MS2Schema>
 {
     private static final Set<String> HIDDEN_PROTEIN_GROUP_COLUMN_NAMES = Collections.unmodifiableSet(new CaseInsensitiveHashSet(Arrays.asList("RowId", "GroupNumber", "IndistinguishableCollectionId", "Deleted", "HasPeptideProphet")));
@@ -388,7 +384,7 @@ public class ProteinGroupTableInfo extends FilteredTable<MS2Schema>
     
     public void addProteinNameFilter(String identifier, MatchCriteria matchCriteria)
     {
-        List<String> params = MS2Controller.getIdentifierParameters(identifier);
+        List<String> params = SequencesTableInfo.getIdentifierParameters(identifier);
         SQLFragment sql = new SQLFragment();
         sql.append("RowId IN (\n");
         sql.append("SELECT ProteinGroupId FROM ");
