+++ protein/api-src/org/labkey/api/protein/PeptideFilterSearchForm.java
@@ -1,0 +1,10 @@
+package org.labkey.api.protein;
+
+public class PeptideFilterSearchForm extends PeptideSearchForm
+{
+    @Override
+    public PeptideSequenceFilter createFilter(String sequenceColumnName)
+    {
+        return new PeptideSequenceFilter(getPepSeq(), isExact(), sequenceColumnName, null);
+    }
+}
