+++ protein/src/org/labkey/protein/view/PepSearchView.jsp
@@ -18,7 +18,7 @@
 <%@ page import="org.labkey.api.protein.PeptideSearchForm" %>
 <%@ page import="org.labkey.api.view.HttpView" %>
 <%@ page import="org.labkey.api.view.JspView" %>
-<%@ page import="org.labkey.ms2.PepSearchModel" %>
+<%@ page import="org.labkey.api.protein.PepSearchModel" %>
 <%@ taglib prefix="labkey" uri="http://www.labkey.org/taglib" %>
 <%@ page extends="org.labkey.api.jsp.JspBase" %>
 <%
